refactor(stella-cli): split the PR-observation helpers out of command_deck.rs
Phase 3's re-query wiring added 6 lines to command_deck.rs, a grandfathered
god file closed to growth, and the file-size ratchet failed PR #3357 at
4103 lines against its 4097 ceiling.

Raising the baseline to clear it is the expedient CLAUDE.md forbids, so the
room comes from a split instead: pr_status_token, ci_status_token, gh_json,
scrub_gh_command, observe_pr and aggregate_ci move to a new sibling module
command_deck/pr_observe.rs. They are the largest self-contained cluster in
the file — pure folds plus single gh invocations — and spawn_pr_monitor
stays behind because it owns the deck's channels.

command_deck.rs is 4011 lines; check-file-size.sh passes. The baseline is
untouched, so the ceiling does not move.
