+++ crates/stella-cli/src/command_deck/tests.rs
@@ -1,3 +1,4 @@
+use super::pr_observe::scrub_gh_command;
 use super::skills::{
     build_skill_creation_prompt, extract_skill_md, extract_skill_md_from_use, parse_installs_count,
     parse_skill_hits, rank_hits,
